LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-beyond-cookies.SharedWorker.sub.https.window-expected.txt


FAIL Verify StorageAccessAPIBeyondCookies for Shared Worker assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for SharedWorker" but got "Unable to load handle in cross-site context for all"
